Prepare utility component migration
