+++ RoseLynn.Generators.VS2019/IGeneratorExecutionContext.cs
@@ -67,4 +67,3 @@ public static class IGeneratorExecutionContextExtensions
         return new(context);
     }
 }
-
